backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-multipart==0.0.20
Pillow==11.3.0
moviepy==2.2.1
imageio-ffmpeg==0.6.0
numpy==2.2.6
pytest==8.4.1
httpx==0.28.1
